+++ genyakubu-manager/src/App.jsx
@@ -1013,6 +1013,7 @@ export default function App() {
               biweeklyAnchors={biweeklyAnchors}
               sessionOverrides={sessionOverrides}
               daySchedules={daySchedules}
+              adjustments={adjustments}
               isAdmin={isAdmin}
             />
           )}
@@ -1277,6 +1278,7 @@ export default function App() {
               examPeriods={examPeriods}
               displayCutoff={displayCutoff}
               daySchedules={daySchedules}
+              adjustments={adjustments}
               onAddStaff={staffCrud.addStaff}
               onDelStaff={staffCrud.delStaff}
               onToggleStaffSubject={staffCrud.toggleStaffSubject}